Adjust README (#76)
The README previously suggested that 'clean' code is a necessity for
amenability to verification. This was not intentional, and as noted
in #63, not true.

This commit removes the mentioning of 'clean' code from the README.
The primary qualities that we seek from 'cleanliness' are amenability
for manual audit and adaptability. The former was already mentioned
explicitly in the README, while the latter is added in this commit.

Closes #63

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>