Keccak: Remove mix of XOF vs. SHAKE128
The reference implementation hides the specific choice of XOF
behind a shallow internal interface resolving to shake128.

The batched implementation of gen_matrix bypasses this interface
by directly calling into shake128x4, while still using the shim
xof-api for non-batched invocations.

This distinction is confusing: We should either use a XOF wrapper
throughout -- both for batched and non-batched calls -- or not use
it at all.

This commit removes the XOF wrapper for now.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>