Hide implementation details of batched SHAKE state
The internal structure of the batched SHAKE state is irrelevant
to the caller; it only needs to know its size in order to be able
to allocate it on the stack.

This commit makes keccakx4_state an implementation detail of
fips202x4.c, and instead exposes shakex4_state as a raw byte
array.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>