Add cache for scaled polys/polyvecs in NTT domain
Asymmetric multiplication [1] speeds up matrix-vector
multiplication in NTT-domain via additional caching
on the polyvec argument.

This commit provides a C implementation of this strategy.

It introduces types poly_mulcache and polyvec_mulcache
for precomputed data used by poly/polyvec base multiplication.
Those types are meant to be opaque to the caller, an exposed
only to allow their construction on the stack.

The mulcache for a poly/polyvec can be computed via
`poly[vec]_mulcache_compute()`. It is then passe to
base multiplication routines as an additional argument.

[1]: https://eprint.iacr.org/2021/986
     Neon NTT: Faster Dilithium, Kyber, and Saber on Cortex-A72 and Apple M1

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>