Fix gen_matrix()
When building the public matrix, we first squeeze GEN_MATRIX_NBLOCKS
out of the Keccak state and rejection sample them into the matrix
entries. We run a 'fill up' stage, squeezing and sampling one block a time,
until the entry buffers are full.

The previous code got the fill-up stage wrong: Instead of squeezing one block
a time, we squeezed GEN_MATRIX_NBLOCKS, _but_ sampled as if we had only squeezed
one. This would lead to incorrect results if the fill-up stage needs more than
one iteration. Thanks to the large default value of GEN_MATRIX_NBLOCKS, this
would not happen in any of the KAT tests, but it may still fail in rare cases.
It also fails when setting GEN_MATRIX_NBLOCKS=2.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>