Fix lint script
The `clang-format` step in the `lint` script would correctly fail
upon badly formatting source files, but (a) the error would not be
displayed correctly, (b) fail to propagate to the result of  `lint`.

This commit fixes this by adjusting the output of the format checker
to match the expectation of `lint`'s internals:
- The output must be to stdout
- The output must consist of lines of the form "file:line:..."
  showing every ill-formatted line/file combination once
The output of `clang-format` does not match this format,
and thusneeds a bit of massaging.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>