Simplify C<->Native interface for mulcache computation
Context: The mul[tiplication]cache of a polynomial is an opaque
structure that implementations can use to cache precomputations
needed during the repeated base multiplication of a polynomial
in NTT form.

Previously, the API for the computation of the mulcache would
expose two arrays of constants which the AArch64 native implementation
of the mulcache computation need.

This commit removes those constants from the native API for the
mulcache computation, since different native implementations are
likely to require different sets of constants.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>