Reduce pk & sk after unpacking
It's not standards-required for the unpacked sk to be be reduced
(its coefficients can be up to 2^12 in size), but our bounds
reasoning assumes it to be.

For pk, it must be checked at the top-level that the byte stream
unpacks to coefficients in bound, which however has not yet been
implemented. Until that's done, reduce the pk after unpacking, so
that the lower level routines are only called with canonical pk.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>