Introduce single contractual bound for [inv]NTT output
Reasoning about safety of the C 'frontend' should not depend
on details of the arithmetic backend (ref/native). We thus
introduce single bounds NTT_BOUND and INVNTT_BOUND on the
absolute values of the output of the [inv]NTT that any
implementation needs to satisfy.

For every specific implementation, we still define and check
(in debug builds) for tighter bounds, plus add a static assertion
that the implementation-specific bound is smaller than the
contractual bound.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>