Document and runtime-check bounds for arithmetic data
This commit documents bounds on arithmetic data as it flows
through common operations like NTT, invNTT and basemul.

Moreover, it introduces debug macros like POLY_BOUND or
POLYVEC_BOUND which can be used in debug builds to check
the documented bounds at runtime. Moreover, at a later
point those assertions should be proved by converting them
to CBMC assertions.

Finally, the bounds show that canonicity of the result of
the NTT is not needed, and this commit removes the Barrett
reduction at the end of the AArch64 ASM NTT.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>