Change semantics of poly_reduce() to unsigned canonical output
Previously, the post-condition for `poly_reduce()` was that its
output coefficients are _signed_ canonical, an unsuitable choice
in two ways:

- `poly_reduce()` is only used in the context of (de)serialization,
  where we're dealing with _unsigned_ canonical data. This mean that
  even though we were calling `poly_reduce()` ahead of `poly_tobytes()`,
  the latter did still need to do a conditional addition.
- The AVX2 implenentation of `poly_reduce()` does not produce signed
  canonical representatives, but 'almost' unsigned canonical ones in
  [0,1,...,KYBER_Q] (inclusive upper bound!) that are then subject to
  a conditional _subtraction_ during deserialization.

This commit is an attempt to clear up the situation by requiring
that `poly_reduce()` produce _unsigned canonical_ representatives,
and adjusting the C and native AArch64 implementations as well as the
serialization routines accordingly.

This paves the way for the integration of AVX2 assembly for poly_reduce().

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>