Remove modular reduction after AVX2 invNTT and NTT
Previously, the AVX2 NTT and invNTT assembly was suffixed
by a call to `poly_reduce()` to ensure polynomial coefficients
sufficiently small that the subsequent arithmetic would not
overflow.

If we trust the bounds for the output of the NTT and invNTT provided
by the Kyber repository, those reductions are not necessary: The bounds
are already below the contractual bounds for the C<->native interface
on the basis of which correctness of the remaining code can be argued.

This commit therefore removes the reductions after the AVX2 NTT and invNTT.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>