Smaller performance improvements of AVX2 backend (#221)
* Remove modular reduction after AVX2 invNTT and NTT

Previously, the AVX2 NTT and invNTT assembly was suffixed
by a call to `poly_reduce()` to ensure polynomial coefficients
sufficiently small that the subsequent arithmetic would not
overflow.

If we trust the bounds for the output of the NTT and invNTT provided
by the Kyber repository, those reductions are not necessary: The bounds
are already below the contractual bounds for the C<->native interface
on the basis of which correctness of the remaining code can be argued.

This commit therefore removes the reductions after the AVX2 NTT and invNTT.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* Add AVX2 implementation of `poly_tomont()`

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* bench_kyber: Print percentiles in addition to median

This commit modifies bench_kyber to not only print the median
execution time of keygen,encaps and decaps, but also some
percentiles. This provides a measure for the stability of the
performance measurements.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* Bump the number of iterations and tests in bench binary

Every data point is the average of X consecutive runs of
the test functions, and we acquire Y data points. This
commit bumps Y from 200 to 500.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* Add AVX2 version of poly_add()

This commit modifies the x86_64 native backend to use AVX2 intrinsics
for the accumulation of data during the computation of a vector-vector
product.

As with the rest of the x86_64 backend, the implementation is taken from
https://github.com/pq-crystals/kyber/blob/main/avx2.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

---------

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>