Absorb zip/unzip of Keccak-x4 state into AVX2 native implementation
The FIPS202 C<->native interface operates on de-interleaved batched Keccak
states. That is, the Keccak-x4 state is the concatenation, not the
interleaving, of four individual Keccak states.

In contrast, SIMD implementations of batched Keccak naturally operate on
interleaved states. This is in particular true for the AVX2 implementation
from XKCP which we are presently using. To remedy, the native interface
implementation conjugates the XKCP core with zipping/unzipping functions
for the (de)interleaving of Keccak states.

This commit slightly improves performance by absorbing the zip/unzip of
Keccak-x4 states into the XKCP implementation of Keccak-x4. Specifically,
the initial memory->register loads are replace by gather loads, and the
final register->memory stores are replaced by gather stores. Note that
while AVX2 has a gather load, it does not have a gather store (that's
introduced in AVX512), so one has to manually transpose the data.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>