Don't generate librng.a libnistrng.a
Testing uses two RNGs that are not meant to be used in production.
Currently, however, those RNGs are still exported as `.a` files of
generic names (librng.a and libnistrng.a) not hinting at the test-only
scope. This increases the risk of users linking them against a production
binary, thereby undermining the security of their system.

To reduce this risk, this commit removes librng.a and libnistrng.a,
and instead adds the relevant source files directly to the test
targets for which they are needed.

The `randombytes.h` header is moved back into `mlkem/*`.

Fixes #286

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>