Remove outdated comment about poly_frombytes()
Previously, `poly_tobytes()` included normalization from signed
to unsigned coefficients. In consequence, it was not an inverse
to `poly_frombytes()`, which was confusing enough to be remarked
in poly.h.

In the interim, `poly_tobytes()` has changed to no longer normalize
from signed to unsigned coefficients, but _assume_ unsigned coefficients
instead. The comment regarding poly_tobytes() and poly_frombytes() not
being inverse is thus outdated, and this commit removes it.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>