Force twiddle as second argument to `fqmul()`
For bounds reasoning, we need to know that `fqmul()` is
always called with a signed canonical value in one of its
arguments.

To simplify the spec, this commit reorders arguments to `fqmul()`
so that the signed canonical twiddle is always the _second_ argument.

It checks this through a debug-only assertion, and also adds
a debug-assertion that the return value of `fqmul()` is always
of absolute value < MLKEM_Q.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>