Debug: Fix documentation of `mlkem_debug_check_bounds()`
`mlkem_debug_check_bounds()` and its wrapper macros `BOUNDS()`
etc. are working with _exclusive_ bounds, but the documentation
for `mlkem_debug_check_bounds()` wrongly stated that its bounds
arguments would be inclusive.

This commit fixes the documentation of `mlkem_debug_check_bounds()`
to say that bounds are exclusive.

No change to the code is needed: We only work with the wrapper
macros, and those were already documented and used with exclusive
bounds.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>