Add input validation for public key and secret key
FIPS204 mandates input validation of the public key in encapsulation,
and for the secret key in decapsulation.
This commit adds the input validation for both.

Input validation for the public key is described in Section 7.2 and ensures
that coefficients in the public key are in the valid range [0, q-1].
We check this by decoding the public key, reducing to the canonical unsigned
representative, re-encoding the reduced polynomial, and comparing it to
the input public key.

Input validation for the secret key is described in Section 7.3 and ensures
that the hash of the public key (that is part of the secret key) actually
corresponds to the real hash of the public key.
We check this by re-computing the hash.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>