Avoid non-integral quotient in documentation of fqmul()
fqmul() is documented to require a 2nd input that's "< q/2"
in absolute value. This is the correct meaning when reading
q/2 as a _rational_ quotient, but it's not when q/2 is read
as round-down division: In that case, one would need to say
<= instead of <.

This commit avoids any confusion by rewriting the documentation
in terms of the integral quotient (q+1)/2.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>