CBMC: Add spec + proof for `poly_add()`
This commit adds a CBMC spec and proof for `poly_add()`.

Previously, poly_add() had the signature (r[out], a[in], b[in]), allowing
the output polynomial to be separate from the two input polynomials.
Yet, uses of `poly_add()` always have the form `poly_add(r,r,b)`,
aliasing the first and second parameters, and thereby accumulating `b`
onto `r`.

Since aliasing poses additional verification challenges, and the
non-aliased version of `poly_add()` is never used, this commit firstly
modifies the signature of `poly_add()` to only take two inputs,
accumulating the second onto the first.

With that done, the spec and proof in CBMC are straightforward, taking
care to use `OLD(...)` and `LOOP_ENTY(...)` in spec and loop invariants
to refer to the original values of memory.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>