KeyGen: Hoist core of matrix entry generation into helper functions
Previously, the matrix generation was done in a single routine, gen_matrix.
This routine is among the most complex in MLKEM. In particular, verification
is likely going to be challenging with the monolithic structure and up to
three nested loops.

This commit aims to facilitate readability and verifiability by hoisting
the core of the matrix generation into separate helper functions:

- gen_matrix_entry(), generating a single matrix entry from a seed buffer
- gen_matrix_entry_x4(), generating 4 matrix entries from 4 seed buffers

Apart from splitting the code between gen_matrix and the new helper
functions, some more comments have been added.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>