CBMC: Pass uninitialized pointers in all verification harnesses
Previously, when a function under proof would assume a pointer to
a structure, the verification harness would construct an instance
of that structure on the stack and pass its address.

This approach is unsound, because it adds to the contractual assumptions
of the function the specifics of the function invocation in the test
harness. For example, a missing bounds constraint in the spec might
go undetected just because the stack-allocated variable happens to
satisfy it.

The most robust way to deal with this, so it seems, is to minimize
the harnesses to merely pass uninitialized variables of the right type
to the function under proof. In particular, where a pointer is expected,
pass an uninitialized pointer, rather than the address of a stack allocated
structure.

Also, remove redundant `x != NULL` preconditions when `IS_FRESH(x,...)`
is assumed. This is not only redundant, but also error prone because of
https://github.com/diffblue/cbmc/issues/8492.

This commit implements those changes for all harnesses implemented so far.

As it turns out, a few specs were incorrect because of missing IS_FRESH
annotations, hidden by the correct allocation in the harness.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>