Add spec + proof for poly_sub and change poly_sub to to argument form
This first changes poly_sub from the 3 argument form poly_sub(r, a, b)
to a two argument form poly_sub(r, b) assuming that r=a.
This makes the proof possible as we do not have to reason about aliasing.
Unfortunately, currently we have poly_sub(x, a, x) rather than
poly_sub(x, x, a) in the single call site in indcpa_dec.
However, it is easy to change it there.

Resolves #308

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>