Modify signature of rej_uniform to work on whole buffer
This commit changes the signature of the rejection sampling
routine rej_uniform() to make its specification more usable
in the context of verification with CBMC.

Previously, rej_uniform would expect a buffer and a target
number of coefficients to be sampled. The caller would then
repeatedly gather fresh input bytes and call rej_uniform,
passing as the buffer base the target buffer _shifted by
the amount of data that has already been sampled_.

This reasoning about passing the shifted target buffer into
rej_uniform seems to be difficult for CBMC -- for reasons I
do not yet understand.

To remedy, this commit modifies the signature of rej_uniform()
so that it always expects the original buffer base, the original
target, and the current offset. It will then sample more
data and return the new offset.

This change does not harden verification of rej_uniform itself,
but will make the verifiation of gen_matrix() much simpler.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>