Improve readability of scalar compression by hoisting pragma
The scalar compression routines include a multiplication in uint32_t
which can overflow. This is expected, and a pragma is used to disable
the default overflow check in CMBC. However, the pragma and the surrounding
ifdefs makes the routine difficult to read.

This commit moves the preprocessor directives related to this outside
of the scalar compression routines. While this unnecessarily widens the
scope of the disabled overflow check, there is no other operation in the
code that would be prone to overflow.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>