Remove redundant conditional addition in polyvec_compress
Contrary to the Kyber reference implementation, poly_reduce() now
guarantees _unsigned_ canonical representatives. Since the input
of `polyvec_compress()` comes out of `poly_reduce()`, the unconditional
addition of `MLKEM_Q` in that routine is therefore no longer necessary.
This commit removes it, and also adds a runtime assertion checking that
indeed the input is unsigned canonical. It also adds an assertion that,
similarly, the output of polyvec_decompress is unsigned canonical.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>