Rename scalar_[de]compress_q_[16,32] -> scalar_[de]compress_d[4,5]
The naming of scalar_[de]compress_q_16_32 is confusing because the
'16' and '32' wrongly suggest a relation to the input/output width.
Instead, those are the compression/decompression routines from
FIPS 203, Eq. 4.7 and Eq 4.8, with d=4 and d=5.

Further, there are (de)compression routines for other values of d
in use, which are currently not yet hoisted into scalar helper
functions.

To avoid confusion and pave the way for more values of d, this
commit renames scalar_[de]compress_q_{16,32} to scalar_compress_d_{4,5}.

The commit also removes the irrelevant suffix `_16` from
`scalar_signed_to_unsigned_q_16`, renaming it to
`scalar_signed_to_unsigned_q`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>