nix: refactor flake.nix for readability and easier testing
Refactoring some let binging into nix packages make it easier for
testing.

For example, we can now test the CBMC dependencies by running, `nix
build .#cbmc`, and check the built dependencies in `results/bin/`

The `litani` derivation is therefore needed to be adjusted so that the
python runtime dependencies can be found.

Signed-off-by: Thing-han, Lim <15379156+potsrevennil@users.noreply.github.com>