KeyGen: Avoid passing an array of poly pointers
The A-matrix generation routine gen_matrix() first generates 4
matrix entries a time, and then does one more, if necessary.
(For MLKEM_K=2 and MLKEM_K=4, the number of matrix entries is
a multiple of 4, while for MLKEM_K=3, we need 9=2*4+1 matrix
entries).

The function taking care of generating 4 entries a time is
gen_matrix_entry_x4(). Previously, this function was rather
generic in that it took an array of 4 pointers to polynomials
as input, allowing them to be scattered in memory.

This generality is both unneeded and problematic for verification:
- It's unneeded, because we always generate a _slice_ of the matrix
  at a time -- that is, the pointers passed are always consecutive.
- It's problematic for verification in CBMC, since I am unaware of
  a way to specify a precondition saying that the 4 given polynomials
  may be part of the same parent object, but must be disjoint in it.

This commit modifies the signature of gen_matrix_entry_x4() to just
pass the address of the sizeof(poly)*4-sized slice of the matrix to
be generated as an argument. This slightly simplifies the caller, since
we don't need to prepare an array of pointers. It also removes the
issue for CBMC verification, since now gen_matrix_entry_x4() only
needs to specify a single validity condition for the entire slice.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>