KeyGen: Use separate stack-allocated seed buffers
Currently, the A-generation routine gen_matrix() uses
a single 2D seed array to manage the 4 seed buffers needed
for the batched generation of 4 matrix entries at a time.

Further down the call stack, slices of that same seed buffer
are passed to the shake256x4-API for absorbing. As in the
previous commit, this causes issues with verification, since
CBMC does not support -- or, at least I am unaware of --
specifying disjointness of slices within the same object.

This commit avoids the issue by allocating separate seed
buffers on the stack, rather than a single 2D array.
This also allows to avoid the manual padding of the seed
buffer, and use standard alignment constraints instead.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>