CBMC: Add spec + proof for gen_matrix_entry_x4
This commit adds a CBMC sped + proof for gen_matrix_entry_x4(),
the part of the A-matrix expanion which generates a slice of four
matrix entries at a time.

While the core functions rej_uniform and shake128x4_xxx are
axiomatized or specified, the proof still proved to be rather
fiddy to get through.
- One has to be careful to spell out the 0,1,2,3 cases, rather than
  using static loops. I don't know why, but somehow the proof gets
  much slower otherwise (in fact, I don't know if it'll terminate
  in acceptable time).
- As in the previous commit, one has to split the temporary sampling
  buffer from a 2D array into four separate stack buffers, to avoid
  (seeming) limitations around pointers to the same object in CBMC.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>