KeyGen: Remove size annotation from gen_matrix_entry_x4()
gen_matrix_entry_x4() expects a slice of 4 polynomials to be filled.
In the caller gen_matrix(), this slice is constructed as a pointer
into an array of MLKEM_K polyvecs. Importantly, however, the slice of
4 poly's passed to gen_matrix_entry_x4() can cross the boundary
between two polyvec instances. Some compilers notice that from
the signature of gen_matrix_entry_x4(), and fail.

The behaviour of the code is intentional and safe here. To suppress
the compiler warning, this commit changes the signature of
gen_matrix_entry_x4() to use a generic poly* pointer, rather than
a pointer to a poly[4].

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>