CBMC: Add spec + proof for polyvec_compress()
This commit adds a largely straightforward spec + proof for
polyvec_compress(). Care needs to be taken to track the bounds
of the compressed data, and to introduce explicit masking when
writing out the individual bytes, since in the configuration used
by our CI, CBMC fails upon implicit truncation that cannot be
proved to be effectless.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>
Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>