Hoist matrix-vector multiplication into separate function
Attempts to verify the toplevel MLKEM-INDCPA API have showed
that the matrix-vector multiplication embedded in the keypair
generation (and, likely, also the one in the encapsulation)
are causing performance issues for CBMC verification.

This commit hoists the matrix-vector multiplication loop,
including the preceeding mulcache computation, into a
helper function `matvec_mul`. This helper function is specified
and proved independently.

To make this work, the call to `poly_tomont()` in the matvec
multiplication loop inside key generation needs to be hoisted
out of the loop and replaced by a call to the new function
`polyvec_tomont()`, which is also given the obvious spec+proof.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>