Strengthen assigns clause in polyvec_basemul_acc_montgomery_cached
Previously, the spec for polyvec_basemul_acc_montgomery_cached()
would mark the entire destination object around the target
polynomial as tainted. This is too strong if the caller passes
a polynomial which is part of a polynomial vector -- which is
the case during matrix vector multiplication.

This commit strengthens the assignment clause for
polyvec_basemul_acc_montgomery_cached to mark only a poly-sized
slice as tainted.

Similarly, the specification of poly_tomont() is strengthened.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>