Allow various object-sharing patterns in poly_getnoise_eta1_4x
Previously, poly_getnoise_eta1_4x() was specified as
requiring four output polynomials satisfying IS_FRESH(...).
While intuitively we only want to express that the output polynomials
are disjoint, the obligation generated for a caller is, in fact, much
stronger: The target polynomials cannot belong to the same object.

In particular, the current specification would not allow to pass
in four consecutive slices of a polyvec, or say 2 consecutive polyvecs
in the first and last two arguments each.

Those limitations will prevent verification of indcpa_keypair_derand,
which does indeed pass slices of polyvecs to the getnoise functions.

To accommodate, this commit introduces multiple preconditions
varying on in the amount of object-sharing that's permitted:
For MLKEM_K=4, the output polynomials must be consecutive slices
of the same object. For MLKEM_K=2, the first and last two must be
consecutive, but belong to separate objects. For MLKEM_K=3, the
first three must be consecutive, and the forth is unused and must
be NULL.

For MLKEM_K=3, which needs sampling of 6 polynomials during keygen,
the code previously used 2 dummy buffers for the two unused output
slots. This commit changes this behaviour and instead passes NULL
as the output buffer, and drops the respective write. This is both
cleaner and faster.

Object sharing in the sibling functions poly_getnoise_eta2_4x
and poly_getnoise_eta1122_4x will be addressed once they are
needed.

One can attempt to use a single REQUIRES(...) clause with a large
disjunction, which proves OK but seems to cause issues at call-time;
this needs further investigation.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>