Allow various object-sharing patterns in poly_getnoise_xxx_4x
Previously, poly_getnoise_eta{1,2,1122}_4x() were specified as
requiring four output polynomials satisfying IS_FRESH(...).
While intuitively we only want to express that the output polynomials
are disjoint, the obligation generated for a caller is, in fact, much
stronger: The target polynomials cannot belong to the same object.

In particular, the current specification would not allow to pass
in four consecutive slices of a polyvec, or say 2 consecutive polyvecs
in the first and last two arguments each.

Those limitations will prevent verification of indcpa_keypair_derand,
which does indeed pass slices of polyvecs to the getnoise functions.

To accommodate, this commit modifies the preconditions of the
poly_getnoise_etaX_4x functions to allow the various object-sharing
patterns the code encounters: Depending on MLKEM_K, the polynomials
are all consecutive in the same object, or split in 2-2, or in 3-1.

Interestingly, and perhaps surprisingly, CBMC handles this disjunction
very well and without significant proof slowdown.

Note: It appeared important to use IS_FRESH() only _after_ specifying
the displacement constraints on the pointers.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>